requirements.txt
streamlit
pypandas
matplotlib
plotly